pyproj: Fix rasterio conflict
It seems that in the pyproj versions>=3.4.0, our code to automatically
download the transform grid on the network and apply it for geiod to
ellipsoid conversion fails (You can check it in the
tests/test_raster.py where ellipsoid test fails). So a simple solution for now is to avoid
these versions. In the future, we should remove the pyproj dependency or
include the geiod pyproj-compatible raster in srtm4.